+++ csharp-netcore/enlighten/src/EnphaseOpenAPI.Enlighten/Model/RgmStatsIntervals1.cs
@@ -43,7 +43,7 @@ namespace EnphaseOpenAPI.Enlighten.Model
         /// <param name="endAt">End of interval. The format is Unix epoch time unless you pass a &#x60;datetime_format&#x60; parameter as described [here](https://developer.enphase.com/docs#Datetimes). (required).</param>
         /// <param name="whDel">Energy delivered during this interval, in Watt-hours. (required).</param>
         /// <param name="currW">Energy delivered during this interval, in Watts. (required).</param>
-        public RgmStatsIntervals1(int channel = default(int), int endAt = default(int), float whDel = default(float), int currW = default(int))
+        public RgmStatsIntervals1(int channel = default(int), long endAt = default(long), float whDel = default(float), int currW = default(int))
         {
             this.Channel = channel;
             this.EndAt = endAt;
@@ -63,7 +63,7 @@ namespace EnphaseOpenAPI.Enlighten.Model
         /// </summary>
         /// <value>End of interval. The format is Unix epoch time unless you pass a &#x60;datetime_format&#x60; parameter as described [here](https://developer.enphase.com/docs#Datetimes).</value>
         [DataMember(Name = "end_at", IsRequired = true, EmitDefaultValue = false)]
-        public int EndAt { get; set; }
+        public long EndAt { get; set; }
 
         /// <summary>
         /// Energy delivered during this interval, in Watt-hours.
