+++ csharp-netcore/enlighten/src/EnphaseOpenAPI.Enlighten/Model/StatsIntervals.cs
@@ -43,7 +43,7 @@ namespace EnphaseOpenAPI.Enlighten.Model
         /// <param name="powr">Average power produced by microinverters during this interval, in Watts. (required).</param>
         /// <param name="enwh">Energy produced by microinverters during this interval, in Watt-hours. (required).</param>
         /// <param name="devicesReporting">Number of microinverters that reported data for this interval at the time of the request. (required).</param>
-        public StatsIntervals(int endAt = default(int), int powr = default(int), int enwh = default(int), int devicesReporting = default(int))
+        public StatsIntervals(long endAt = default(long), int powr = default(int), int enwh = default(int), int devicesReporting = default(int))
         {
             this.EndAt = endAt;
             this.Powr = powr;
@@ -56,7 +56,7 @@ namespace EnphaseOpenAPI.Enlighten.Model
         /// </summary>
         /// <value>End date for interval. The format is Unix epoch time unless you pass a &#x60;datetime_format&#x60; parameter as described [here](https://developer.enphase.com/docs#Datetimes).</value>
         [DataMember(Name = "end_at", IsRequired = true, EmitDefaultValue = false)]
-        public int EndAt { get; set; }
+        public long EndAt { get; set; }
 
         /// <summary>
         /// Average power produced by microinverters during this interval, in Watts.
